Move cli commands to respective agent, controller and common directories
Also removes the wrappers around the cli commands, as these are
no longer used. The only agent command is a new one - kill_task, which
didn't have a wrapper anyway, and the controller commands are now
wrapped in django manage commands so they can be called more easily
in the dokku app.