Allow Django ORM to read the controller database.
This is a proof-of-concept commit/PR.

Change the `default` database (which was previously only used for testing
purposes) to point at the default location for the controller database. This
allows opening a connection, `manage dbshell` etc.

If we were doing this for real probably we would read the DB location from some
configuration but here it's hardcoded.

This also allows `manage inspectdb` to generate the models.py seen here. These
have been edited to remove all the `blank=True null=True` added by default, to
make the primary keys valid. This enables importing and using the models in
modules or eg via `manage shell`.