Reimplement `calculate_workspace_state` with queryset.
This is a proof-of-concept commit and PR.

In a management command, reimplement `calculate_workspace_state` using
querysets and a partition approach and compare perfomance with the extant
implementation that uses our homebrew ORM and some manipulation in Python.

Very brief initial profiling shows this as approximately 50 times faster on my
machine when tested against around 2000 jobs/action from the
https://github.com/opensafely/disparities-comparison repo.

This may demonstrate that this function can be optimized sufficiently to use
synchronously, and that using the Django ORM is a viable alternative to
optimizing the homebrew ORM code used. The Django ORM approach can be used
alonside code using the homebrew approach, which could enable a gradual
switchover.

Clearly the homebrew ORM code could be optimized further (eg it does some work
in Python that could be off-loaded to the database engine with `FIRST_VALUE`
and `PARTITION`, and I think we know the value-type unpacking is slow). Maybe
there's a better queryset approach.

My very brief assessment is that the queryset code needs at most three passes
through the rows, O(n), where the extant approach uses a sub-group and sorting
approach that may be O(n log n) or possibly O(n^2 log n).