Add initial configuration and plumbing for restricted datasets
The initial permissions configuration is taken from the current yaml
config in opensafely cli, with a number of changes made to match the way
t1oo's work in job-server, which is where we intended this config to
live long term.

 - key by project rather than repo
 - use project slug rather than number, as job-runner doesn't know that.
 - use python rather than yaml

Have used the names of the tables as defined in opensafely-cli for now,
we may want to change the string values once we know what they will be.

The plumbing is fairly straightforward, it just sets a job env var with
the values of any permitted tables as a comma separate string for ehrql
to parse.