Refactor the repo checking code into a common function
This will allow us to use the same logic in the agent.

There's a slight change to testing behaviour here. Previously, we
skipped the org test if the repo url did not started with `http`, so we
can pass local git paths in tests and skip that validation.

Here we extend that same skipping logic to the commit/branch
verification too. This removes the need for having a valid repo at all
in the job/task data for tests - it can just be a string.

This reduces complexity and test setup time for test that do not need
a repo to function.

As such, one test that was affected by this needed to use an actual
http repo, rather than a local one.