Use consistant test data values
It was possible to create jobs and job_requests that did not have the
same values for things like workspace, repo, and commit.

This makes them consistant.

Also adds a basic defaults test for job_to_job_definition. I am about to
change it, so wanted a failing test, and it seemed like it was missing.