Add branch to JobDefinition
We need this to be able to check the job's repo/commit/branch in the
agent.

This is a bit awkward, as branch is not on the Job model, its only on
in the original job request json.

We could add it as a column to Job model, however, if we are going to
replace the current ORM with django's, then there's not a lot of point
doing that, as I suspect we'll have a foreign key to the job request
anyway.  And I am reluctant to change Job db schema whilst the RAP API
step 2 work is happening.