Ensure workdir exists in docker tests.
This both ensures that it does, and adds some config to the mount config
so that it will error if it doesn't to avoid this in future.

By default, if workdir does not exist on the host, docker compose will
create it.  Unfortunatlely, it will create it with root ownership, which
messes with the permissions, as the containers do not run as root.

In CI, we run the docker tests in their own job, and it runs as the
runner user, so it doesn't have permissions to write to a root owned
/workdir.

Normally, we use fixtures (db, git checkout stuff) that create tmp dirs,
so stuff has worked w/o issue.

However, recent commits changed that, and in 2 tests, we now *do*
actually checkout stuff into the default configured workdir, which
causes these tests to fail in CI for docker only.

The plan is to change those two tests to do so in isolated tmp dirs as
well, so tests continue not to use the default ./workdir, which is good
idea anyway.

However, before we fix that, we use the fact that this helpfully exposed
this hidden permissions issue, and take steps to make sure it won't bite
us in confusing hard to debug way again in future.