Add `create_jobs` telemetry.
RAP API v2 team are interested in how fast this runs in production in different
cases and if it can be made synchronous.

`create_jobs` should only be called once per new job request, so is not a lot
of extra telemetry.

Breaking out into separate commits to make the diffs a bit easier to follow.