Require ALLOWED_GITHUB_ORGS to be set
It is set by default, but overridable via env var.

However, we used the hack of setting it to None to skip validation in
some tests. This change makes it required, so that we fail closed in
prod if not set. Note that checks are still skipped if the url does not
start with "http", so that we can use local repo file paths in testing.

This is exposed a few places that needed fixing in tests.  I also
changed the error messages to include the url in them, which was helpful
in debugging these test failures.