Config for CLIENT_TOKENS
Add env variables and config for a client token that allows a
client (e.g. job-server) to request/update information for a
discrete list of backends. This allows for endpoints to supply
information about all backends that a client has access to.

Currently the only client is job-server, but in future we may
have multiple clients who are only interested in and allowed to
get/update about a subset of backends.