Use ruamel in yaml utils
This shouldn't make any difference, as ruyamel is a fork of
ruamel.yaml with the same api. We previously replaced ruamel.yaml
with ruyaml (hence the existing comments that refer to ruamel) but
since we're now back to using ruamel for the pipeline library to
use, we might as well switch back to using the explicit job-runner
dependency.