Time more subroutines in `create_jobs` telemetry.
We should time them all in production for long enough to determine which are
'interesting', then keep those.

We're not sure exactly which subroutines account for the few 100ms extra in the
duration of the span in production over local. It's probably
`validate_job_request` and `validate_job_request` as they involve contacting
GitHub which is slower in production than against a local repo checkout.