Replace ruyamel with ruamel in yaml_utils
This shouldn't make any difference, as ruyamel is a fork of
ruamel.yaml with the same api. We previously replaced ruamel.yaml
with ruyaml (hence the existing comments that refer to ruamel) but
since we're now back to using ruamel in the pipeline libary (the
fastparser options), we might as well switch back to using it here too.