Validator helpers for api views
Adds a decorator to be used for validating a RAP api post body.

Expects to be passed a dataclass, with a from_request() classmethod
that will do the actual validating and raise ApiValidationError if it
encounters errors. The decorator will return the error response, or
add the dataclass instance to the view's kwargs as `request_obj`.

The view function should also be decorated with
@get_backends_from_client_token, so we already have backends in the
view args (i.e. the backends that this client token is authorised
to access) and we can validate a backend passed in the request data
against it.

This will allow api views to do their own custom validation, and
provide an object that can be used for whatever operations the view
requires, while removing the boilerplate of validating the request
body and raising errors from the view code itself.