Validator helpers for api views
Adds a decorator to be used for validating a RAP api post body.

Expects to be passed a dataclass, with a from_request() classmethod
that will do the actual validating and raise ApiValidationError if it
encounters errors. The decorator will return the error response, or
add the dataclass instance to the view's kwargs as `request_body`.

This will allow api views to do their own custom validation, and
provide an object that can be used for whatever operations the view
requires, while removing the boilerplate of validating the request
body and raising errors from the view code itself.