Add cancel RAP API endpoint
This is called to request cancellation of one or more actions for a
specific job_request_id.

It currently reuses the set_cancelled_flag_for_actions method and doesn't
do any checks on whether the job_request_id and actions provided
are valid (other than validating the shape of the posted data).