Add cancel RAP API endpoint
This is called to request cancellation of one or more actions for a
specific job_request_id.

It checks that the specified action exist for the specified
job_request_id
but doesn't care whether they've already been cancelled or not.
Uses the existing `set_cancelled_flag_for_actions` method to do the
actual updating.