Refactor auth and validation decorators
Make the view arguments after request keyword only.
The get_backends_for_client_token decorator now adds backends as a
keyword argument, which makes it easier to retrieve in the
validate_request_body decorator, which expects it to have been
applied first.

If we end up in the validate_request_body decorator without a
backends keyword arg, it means a developer has either omitted the
get_backends_for_token decorator, or has applied it in the wrong
order, so raise an AssertionError with a helpful error message.

Also adds specific tests for the decorators and removes the decorator-
imposed auth/validation tests from the rap_views tests