Add `sync_backend` telemetry.
RAP API v2 team are interested in how fast the parts of this run in production
in different cases and in particular to understand if any part is too slow to
run as a cron every minute or if we might need to use `run_once` or another
approach.

Emitting a span every loop would be ~ one every 60 seconds, ~1440 a day ~40k a
month. I can see the Honeycomb plan we're on gives us 100M events per month and
we've emitted 6.4M so far so I estimate about 25M events a month at present. So
I think adding another 40k would be fine and we can always turn it off if it is
a problem.
https://ui.honeycomb.io/teams/bennett-institute-for-applied-data-science/usage