Add server config to openapi spec file
We can specify a relative path as the server url, so we don't
need to prefix every endpoint. If in future we have a v2, we
can add a second server url and apply it to only specific
paths.

Also updates the justfile to pass the urls properly for both the
docker and non-docker versions.