Update the create request body to closer resemble the job request the controller uses
This makes the request body received by the create endpoint contain
(mostly) just what we need for the JobRequest instance we pass around,
in a format that closely matches that.  See
job_request_from_remote_format() in sync.py for the JobRequest format
that the current sync loop generates.

The exceptions are:
- We take a job_request_id which will become the job request's id field.
- cancelled_actions are not relevant for a create request
- project and orgs are required for telemetry and dataset permissions,
  and are retrieved from the SavedJobRequest - in order to maintain the existing code in parallel, we'll need to construct an `original` field that contained
 the fields that are used from the SavedJobRequest. I expect this will
go away later