Make the jsonschema validation errors more informative
If the error happens in the tree somewhere, report the json_path
associated with it - this should have enough information to tell
us something more useful about which part of the request body has
the problem. If we only report the error message, we can end up
with something like "[] is not of type 'object'" for a nested
property in a complex structure, without an indication of which
thing we received the wrong type for.