Create jobs in the create endpoint
Uses the existing create_jobs() code to create jobs; this already
returns a count of jobs, which we now return in the response.

We catch the same known errors as the previous code, as well as
unknown errors, but we don't now create any dummy failed job from
the exception, we just return the error for the client to handle.

create_jobs does some validation on the job request, which isn't
needed in this version of the code, but is still needed while we
have the parallel sync loop using the same code.