Require explicit lists of expected status codes in api spec test
Also adds a test to ensure that the spec is consistent with the
expected status codes we test against. This means we now ignore
405s because they're not included in the spec (and they're not
especially interesting).

Theoretically we could just retrieve the expected status codes from
the spec and use those in the final recorder test, but this way we
can see the paths and codes being tested in the test file.