Stop `sync` cancelling jobs for `test` backend.
Towards RAP API v2. Now that the Job Server UI can be used to cancel `test`
jobs via `webapp`, the synchronisation loop does not need to. This allows us to
test the RAP API cancel without any use of the loop.

If this works well we will do the same for all backends, deleting the code path.