Use JSON to encode ehrQL permissions
This is instead of a comma-separated list. At present permissions are
just lists of strings but we want to leave open the possibility of
passing more complex structured data. Plus, using a "real" encoding
seems better for this purpose than DIYing it, however simple the current
implementation is.

We also include ELD access in the permissions list, as well as setting
the appropriate environment variable. Once ehrQL is updated we can drop
the extra environment variable we gives us a much cleaner API.