Remove macOS tests from CI
We initially had these in place (alongside Windows tests) because the
old codebase needed to work on users' machines. During the
Agent/Controller split work we dropped the Windows tests but kept macOS
because at least one developer was using macOS locally.

This is no longer the case and, due to some current flakiness with
Github's macOS runners, it doesn't seem worth the cost to keep these in
place.

See related Slack thread:
https://bennettoxford.slack.com/archives/C069YDR4NCA/p1759314821692029