Add documentation on how to run multiple backends (agents) at the same time
With a single job-runner repo, we can only simulate running one backend agent.
This adds some docs to describe how to run a second agent, if you want
to test with multiple backend agents calling the same controller.

Also updates the example with a second backend that will actually
work with ehrql jobs (and explains why).