Log less when cancelling multiple actions.
Probably this is Job Server cancelling a whole request.  Let's just log the
number of actions in that case.  Technically partial cancellations are possible
but Job Server will not do that at present, and this avoids spamming 100s of KB
when cancelling large requests.

Add a test for multiple actions to exercise this, though we don't test the
logged values to avoid brittle tests.