Don't cancel jobs in sync loop.
We have tested sufficiently against the test backend in production using the
Job Server route so can disable this in general now.

We should merge https://github.com/opensafely-core/job-server/pull/5376 before
this so that there is a route to cancel any job. It is okay to cancel the same
action multiple times.