Remove openapi-spec-validator dependency
A dependency conflict is causing pip-compile to downgrade
openapi-spec-validator significantly (from 0.7.2 to 0.5.5), to a
version where it doesn't work with our api spec.

Since it's not used in any CI tests, only for local development
when adding to the API spec, and there are alternatives (the
swagger editor and the schemathesis tests), let's just remove it.