Add a daily cron task to take the snapshot
The exact timing of this isn't critical but it should ideally be shortly
before we configure the external snapshot of the volume itself. We
definitely don't want it to happen _during_ the external snapshot, and
having it run sligthly afterwards means that our backups will be 24
hours older than they need to be.