Move `job_request_from_remote_format` to `add_job.py`.
This is the only place it is used. `create_jobs` now takes a
`CreateRequest`, so construct that. Keeping it as a separate
function for now because the `original` dict is still needed.
Private function so implicitly tested by the other module
tests. We may rethink `add_job` as we should not need to
create jobs manually anymore. And `original` may go away.

That was the last thing in the `sync` module, so remove it.