Add some brief notes on how to restore database backups
This took me surprisingly long to work out because you can't do it
through the SnapShooter UI where our backups are configured, but that UI
also gives you no pointers as to where you _can_ do it.
https://bennettoxford.slack.com/archives/C63UXGB8E/p1761043386760259

Once Lucy pointed me in the right direction however it was fairly
straightforward and I've confirmed that it does actually work.