Add library to talk ghrc.io to resolve image shas
This is cribbed from similar code in opensafely-cli. The dockerhub v2
API that ghrc.io users requires an auth token, even for public images,
so we have to handle that, and renewing it when it expires.

For reliability, the get_current_image_sha will use a cached sha on
network failure.

The testing of the ghcr.io interactions is a little complex. Using
passthru and manual mocks for the same url is awkward.