Run images at specific shas.
The controller now resolves the sha of the current image when the runjob
task is created.

In the prepare step, the agent now ensures that specific sha is
available, pulling it if needed, which removes the need for manual
image management altogether.

In the execute step, it runs the image at the specificed sha.

There is however a complication: this is effectively a migration of the
format of task.definition for job tasks.  We will simultaneously have
tasks with and without image_sha data queue.  So, when image_sha is not
set, we fall back to the old way. Once we are sure last old format job
has finished (two weeks?), we can come back and remove the old
codepaths, which will simplify a bunch of things.

The production code is fairly simple, it adds a network call on the
critical path for running jobs that was not there before. This has two
impacts:

Firstly, we don't want to fail jobs if ghcr.io is down. So, we cache
previous shas, and use the stale version in the case of network error.

Secondly, introducing a network call into the many tests that create
runjob tasks is also not desireable. Instead, we add a global autouse
fixture that provices canned shas for the default test usage. We provide
an escape hatch for integration tests, which *do* need the actual shas,
in the form of a `needs_ghcr` pytest marker. This is similar in concept
to `needs_docker`, albeit quite a different implementation. If present,
the autouse fixture will not patch in the canned responses.  The
alternative would have been adding the fixture manually to 80%+ of
tests, which seemed unreasonable.