Use the proxy name to run the actual image
We previously pulled the actual image and sha from the proxy, then locally
tagged it back to use ghcr.io name when running the job.

However, it turns out we cannot use the sha on a locally tagged image,
has to have come from a dockerhub. Under the hood, the original
implementation was reaching out to get the manifest from ghcr.io when we
called `docker run ghcr.io/openafely-core/ehrql@sha256:...`, which
worked in tests but would have failed in production.

Instead, we do not tag it, and run the proxy image name directly. This
is simpler and functionally equivalent.