Tag pulled docker images in a safe manner.
The previous commit uses the DOCKER_PROXY image name to pull and run the
job, which is safest and cleanest.

However, if there are only shas, and no tags, whilst the agent has no
problem running the right image, human operators debugging things on the
backend have no clues as to which sha is what version.

There are times we may manually want to run ehrql or other images
manually, where having tags would be helpful.

So, whlist we still pull and run with DOCKER_PROXY based names, we also
opportunisitically tag them if they are newer than the current version
of that tag. So for example, if we have ehrql:v1@sha1, and we pull down
ehrql:v1@sha2, which has been created *after* sha1, then we will update
the ehrql:v1 to point at sha2.

This should keep the published labels (v1, v2, etc) at the most recent
version. The agent still uses the sha directly, but human operators can
use the semantic tags to run the latest version manually for debugging.