Add a _checkenv recipe to check for .env file in recipes that need it
Using _dotenv or devenv for e.g. test/run doesn't work because the
newly created .env file isn't loaded into the environment, and there
doesn't seem to be an easy way to get just to reload it. Instead, we
just check if it exists and prompt the user to run just devenv.

I think this is a better ux than using _dotenv/devenv as a dependency
and having the command you ran fail but work the next time you
try it. The info message that we printed in _dotenv was getting
hidden behind lots of error text. (test/run-app are OK because we've
got a useful message written into the django settings, but others
e.g. run-agent error confusingly).