Use python3.13 in tomlkit
Our .python-version is 3.10, which is too old for the datetime code used in the bump-uv-cutoff recipe (needs 3.11+). uv should provide the python for us, so we can use a more modern version. All this code does is calculate a new timestamp and write it into the pyproject.toml, so we don't need to care about being consistent with the .python-version here.

Also include upgrade-pipeline in update-dependencies, and follow it with
just devenv instead of prodenv (which uninstalls dev dependencies which
update-dependencies then has to reinstall again)