Ensure all mutations are in a transaction.
To use sqlite concurrently, we *have* to use a transaction for writes,
or else we don't get the good concurrent write semantics we want.
There were some cases where we were not, which likely caused a db lock
in production.

We also use larger, multi-write-statement transactions in some places,
to ensure consistancy, and we don't want get into nested transactions.

So, this change adds an ensure_transaction decorator, and wraps all the
database mutation methods with it.

This ensures we can't accidentally perform a write that is not in
a transaction, but we can still use larger outer transactions to group
related writes together.