Make set_flag manage its own transaction.
This make it much easier to use in various places, and seems reasonable.

To do this, we had to unbundle it in one place it was part of a larger
transaction. However, I think that larger transaction was not needed, as
they are not semantically linked changes.

This is the first wrinkle, where its not clear where the transaction
boundary should be.

But putting a transaction round every single set_flag call is also not
attractive.