Set analysis_scope when creating jobs
We expect analysis_scope to be passed to the RAP API from the client
when creating a job. Eventually, we expect this to contain all the
permissions and access info that a job might need. However, currently
dataset permisisons and access to event level data live in this repo,
so we combine data from both sources to build the analysis_scope for
the job. This also means that we can move to using analysis scope
passed from the RAP API before we remove the permission info from
this repo.