Management command to backpopulate Jobs from SavedJobRequest
This means we can backpopulate all the jobs in the production
database with the new attributes, and ensure that they exist
before we start using them, and before we remove the SavedJobRequest.

This is only intended to be run once, and it logs summary info
deliberately (rather than just writing to stdout) so that info
gets written to the main logs. When we remove the SavedJobRequest
we'll also remove this management command.